cmd: move reporting func to report file
